Implemnt modifying personal info and attendance

+++ project/record/views.py
@@ -9,7 +9,7 @@ from record.models import Post, Question, UserQuestion
 from record.serializers import PostSerializer, QuestionSerializer, UserQuestionSerializer
 from record.filters import DynamicSearchFilter
 from config.permissions import MyIsAuthenticated
-from core.models import UserCoin
+from core.models import UserCoin, Attendance
 from core.serializers import UserCoinSerializer
 
 from django.http import Http404
@@ -121,6 +121,10 @@ class PostCreateView(APIView):
                                 partial=True)
             if uc_serializer.is_valid():
                 uc_serializer.save()
+                Attendance.objects.create(
+                    profile=profile,
+                    date=today
+                )
             
             post = Post.objects.get(profile=profile, created_at=today)
             return Response({
@@ -155,27 +159,70 @@ class PostDetail(APIView):
             raise Http404
 
     def get(self, request, pk, format=None):
+        User = get_user_model()
+        email = request.user.email
+        profile = User.objects.get(email=email)
+
         post = self.get_object(pk)
         serializer = PostSerializer(post)
-        return Response(serializer.data)
+        
+        if post.profile == profile or profile.role == 10:
+            return Response(serializer.data)
+        else:
+            return Response({
+                'response': 'error',
+                'message': '다른 사람의 일기는 볼 수 없어요.'
+            })
     
     def patch(self, request, pk, format=None):
         post = self.get_object(pk)
-        serializer = PostSerializer(post, data=request.data, partial=True)
-        if serializer.is_valid():
-            serializer.save()
+
+        User = get_user_model()
+        email = request.user.email
+        profile = User.objects.get(email=email)
+
+        if post.profile == profile or profile.role == 10:
+            serializer = PostSerializer(post, data=request.data, partial=True)
+            if serializer.is_valid():
+                serializer.save()
+                return Response({
+                    "response": "success", 
+                    "message": "성공적으로 수정하였습니다."})
             return Response({
-                "response": "success", 
-                "message": "성공적으로 수정하였습니다."})
-        return Response({
-            "response": "error", 
-            "message" : serializer.errors
-            }, status=status.HTTP_400_BAD_REQUEST)
+                "response": "error", 
+                "message" : serializer.errors
+                }, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return Response({
+                'response': 'error',
+                'message': '다른 사람의 일기는 수정할 수 없어요.'
+            })
 
     def delete(self, request, pk, format=None):
         post = self.get_object(pk)
-        post.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+
+        User = get_user_model()
+        email = request.user.email
+        profile = User.objects.get(email=email)
+        usercoin = UserCoin.objects.get(profile=profile.id)
+
+        if post.profile == profile or profile.role == 10:
+            if post.created_at == usercoin.last_date:
+                return Response({
+                    'response': 'error',
+                    'message': '마지막 기록은 지울 수 없습니다.'
+                })
+            else:
+                post.delete()
+                return Response({
+                    'response': 'success',
+                    'message': '기록을 삭제하였습니다.'
+                }, status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response({
+                'response': 'error',
+                'message': '다른 사람의 일기는 삭제할 수 없어요.'
+            })
 
 
 class QuestionList(APIView):
